+++ policy/check_test.go
@@ -35,11 +35,11 @@ func (m *mockStrategyRunner) Run(eval *sdk.ScalingCheckEvaluation, count int64) 
 }
 
 type mockAPMLooker struct {
-	t         *testing.T
-	query     string
-	timeRange sdk.TimeRange
-	metrics   sdk.TimestampedMetrics
-	err       error
+	t          *testing.T
+	query      string
+	timeRange  sdk.TimeRange
+	metrics    sdk.TimestampedMetrics
+	err        error
 	queryCalls int
 }
 
@@ -180,14 +180,11 @@ func TestCheckHandler_runAPMQuery(t *testing.T) {
 	mockLooker := &mockAPMLooker{
 		t:     t,
 		query: testQuery,
-		timeRange: sdk.TimeRange{
-			From: now.Add(-testWindow),
-			To:   now,
-		},
 	}
 
 	tests := []struct {
 		name       string
+		instant    bool
 		metrics    sdk.TimestampedMetrics
 		queryError error
 		expResult  sdk.TimestampedMetrics
@@ -195,6 +192,13 @@ func TestCheckHandler_runAPMQuery(t *testing.T) {
 	}{
 		{
 			name:      "valid_metrics_returned",
+			metrics:   testMetrics,
+			expErr:    nil,
+			expResult: testMetrics,
+		},
+		{
+			name:      "valid_metrics_returned_instant_query_window",
+			instant:   true,
 			metrics:   testMetrics,
 			expErr:    nil,
 			expResult: testMetrics,
@@ -216,6 +220,10 @@ func TestCheckHandler_runAPMQuery(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			mockLooker.metrics = tc.metrics
 			mockLooker.err = tc.queryError
+			mockLooker.timeRange = sdk.TimeRange{From: now.Add(-testWindow), To: now}
+			if tc.instant {
+				mockLooker.timeRange = sdk.TimeRange{From: now, To: now}
+			}
 
 			check := &sdk.ScalingPolicyCheck{
 				Name:              "check",
@@ -223,6 +231,7 @@ func TestCheckHandler_runAPMQuery(t *testing.T) {
 				Query:             testQuery,
 				QueryWindow:       testWindow,
 				QueryWindowOffset: 0,
+				QueryInstant:      tc.instant,
 				Strategy: &sdk.ScalingPolicyStrategy{
 					Name: "strategy",
 				},