+++ policy/file/parse.go
@@ -1,4 +1,4 @@
-// Copyright IBM Corp. 2020, 2025
+// Copyright IBM Corp. 2020, 2026
 // SPDX-License-Identifier: MPL-2.0
 
 package file
@@ -70,11 +70,13 @@ func decodePolicyDoc(decodePolicy *sdk.FileDecodeScalingPolicy) error {
 		check := decodePolicy.Doc.Checks[i]
 
 		if check.QueryWindowHCL != "" {
-			w, err := time.ParseDuration(check.QueryWindowHCL)
-			if err != nil {
-				return err
+			if check.QueryWindowHCL != "instant" {
+				w, err := time.ParseDuration(check.QueryWindowHCL)
+				if err != nil {
+					return err
+				}
+				decodePolicy.Doc.Checks[i].QueryWindow = w
 			}
-			decodePolicy.Doc.Checks[i].QueryWindow = w
 		}
 
 		if check.QueryWindowOffsetHCL != "" {